+++ src/app/(route)/(log-in_sign-up)/log-in/page.tsx
@@ -10,9 +10,15 @@ export default function Login() {
           <p className="form-title">{FORM_TEXT[0]}</p>
           <LoginForm />
 
-          <Link href="/change-pwd" className="text-gray-1 underline mt-6">
-            {FORM_TEXT[12]}
-          </Link>
+          <div className="flex mt-6 gap-3">
+            <Link href="/change-pwd" className="text-gray-1 underline mt-6">
+              {FORM_TEXT[12]}
+            </Link>
+            <Link href="/sign-up" className="text-gray-1 underline mt-6">
+              회원가입
+            </Link>
+          </div>
+
           <div className="flex flex-col self-stretch items-center mt-[17px]">
             <span className="mb-4 text-gray-3">{FORM_TEXT[5]}</span>
